feat: add conditional buttons (Modify+Delete / A will do that!); ad Likes SVG.
